Moved rest of `SYSTEM_init` stuff to `stpy`.
