+++ electrical/system/defs.h
@@ -222,198 +222,11 @@ CMSIS_PUT(struct CMSISTuple tuple, u32 value)
 
 
 
-// TODO Document.
-
-/* #meta IMPLEMENT_DRIVER_ALIASES : SYSTEM_DATABASE
-
-    def IMPLEMENT_DRIVER_ALIASES(
-        *,
-        driver_name,
-        cmsis_name,
-        common_name,
-        identifiers,
-        cmsis_tuple_tags,
-    ):
-
-        for target in PER_TARGET():
-
-
-
-            if driver_name not in target.drivers:
-
-                Meta.line(
-                    f'#error Target "{target.name}" cannot use the {driver_name} driver '
-                    f'without first specifying the peripheral instances to have handles for.'
-                )
-
-                continue
-
-
-
-            # Have the user be able to specify a specific driver unit.
-
-            Meta.enums(
-                f'{driver_name}Handle',
-                'u32',
-                (handle for handle, instance in target.drivers[driver_name])
-            )
-
-
-
-            # @/`CMSIS Suffix Dropping`.
-            # e.g:
-            # >
-            # >    I2Cx <-> I2C3
-            # >
-
-            Meta.line(f'#define {common_name}_ {cmsis_name}_')
-
-
-
-            # Create a look-up table to map the moniker
-            # name to the actual underyling value.
-
-            Meta.lut( f'{driver_name}_TABLE', (
-                (
-                    f'{driver_name}Handle_{handle}',
-                    (common_name, instance),
-                    *(
-                        (identifier.format(common_name), identifier.format(instance))
-                        for identifier in identifiers
-                    ),
-                    *(
-                        (
-                            cmsis_tuple_tag.format(common_name),
-                            CMSIS_TUPLE(SYSTEM_DATABASE[target.mcu][cmsis_tuple_tag.format(instance)])
-                        )
-                        for cmsis_tuple_tag in cmsis_tuple_tags
-                    )
-                )
-                for handle, instance in target.drivers[driver_name]
-            ))
-
-
-
-        # Macro to mostly bring stuff in the
-        # look-up table into the local scope.
-
-        Meta.line('#undef _EXPAND_HANDLE')
-
-        with Meta.enter('#define _EXPAND_HANDLE'):
-
-            Meta.line(f'''
-
-                if (!(0 <= handle && handle < {driver_name}Handle_COUNT))
-                {{
-                    panic;
-                }}
-
-                struct {driver_name}Driver* const driver = &_{driver_name}_drivers[handle];
-
-            ''')
-
-            for alias in (common_name, *identifiers, *cmsis_tuple_tags):
-                Meta.line(f'''
-                    auto const {alias.format(common_name)} = {driver_name}_TABLE[handle].{alias.format(common_name)};
-                ''')
-
-*/
-
-
-
 //////////////////////////////////////////////////////////////// System Initialization ////////////////////////////////////////////////////////////////
 
 
 
-// Macros to control the interrupt in NVIC.
-// @/pg 626/tbl B3-8/`Armv7-M`.
-// @/pg 1452/tbl D1.1.10/`Armv8-M`.
-
-#define NVIC_ENABLE(NAME)        ((void) (NVIC->ISER[NVICInterrupt_##NAME / 32] = 1 << (NVICInterrupt_##NAME % 32)))
-#define NVIC_DISABLE(NAME)       ((void) (NVIC->ICER[NVICInterrupt_##NAME / 32] = 1 << (NVICInterrupt_##NAME % 32)))
-#define NVIC_SET_PENDING(NAME)   ((void) (NVIC->ISPR[NVICInterrupt_##NAME / 32] = 1 << (NVICInterrupt_##NAME % 32)))
-#define NVIC_CLEAR_PENDING(NAME) ((void) (NVIC->ICPR[NVICInterrupt_##NAME / 32] = 1 << (NVICInterrupt_##NAME % 32)))
-
-
-
 #include "SYSTEM_init.meta"
-/* #meta
-
-
-
-    # Macros to make using GPIOs in C easy.
-    # TODO Use SYSTEM_DATABASE.
-
-    Meta.define('GPIO_HIGH'  , ('NAME'         ), '((void) (CONCAT(GPIO, _PORT_FOR_GPIO_WRITE(NAME))->BSRR = CONCAT(GPIO_BSRR_BS, _NUMBER_FOR_GPIO_WRITE(NAME))))')
-    Meta.define('GPIO_LOW'   , ('NAME'         ), '((void) (CONCAT(GPIO, _PORT_FOR_GPIO_WRITE(NAME))->BSRR = CONCAT(GPIO_BSRR_BR, _NUMBER_FOR_GPIO_WRITE(NAME))))')
-    Meta.define('GPIO_TOGGLE', ('NAME'         ), '((void) (CONCAT(GPIO, _PORT_FOR_GPIO_WRITE(NAME))->ODR ^= CONCAT(GPIO_ODR_OD , _NUMBER_FOR_GPIO_WRITE(NAME))))')
-    Meta.define('GPIO_SET'   , ('NAME', 'VALUE'), '((void) ((VALUE) ? GPIO_HIGH(NAME) : GPIO_LOW(NAME)))')
-    Meta.define('GPIO_READ'  , ('NAME'         ), '(!!(CONCAT(GPIO, _PORT_FOR_GPIO_READ(NAME))->IDR & CONCAT(GPIO_IDR_ID, _NUMBER_FOR_GPIO_READ(NAME))))')
-
-
-
-    for target in PER_TARGET():
-
-        # For interrupts (used by the target) that
-        # are in NVIC, we create an enumeration so
-        # that the user can only enable those specific
-        # interrupts. Note that some interrupts, like
-        # SysTick, are not a part of NVIC.
-
-        Meta.enums(
-            'NVICInterrupt',
-            'u32',
-            (
-                (interrupt, f'{interrupt}_IRQn')
-                for interrupt, niceness in target.interrupts
-                if INTERRUPTS[target.mcu][interrupt] >= 0
-            )
-        )
-
-
-
-    # Initialize the target's GPIOs, interrupts, clock-tree, etc.
-
-    for target in PER_TARGET():
-
-        with Meta.enter('''
-            extern void
-            SYSTEM_init(void)
-        '''):
-
-
-
-            ################################ Clock-Tree ################################
-
-
-
-            # Figure out the register values relating to the clock-tree.
-
-            configuration, tree = SYSTEM_PARAMETERIZE(target)
-
-
-
-            # Figure out the procedure to set the register values for the clock-tree.
-
-            SYSTEM_CONFIGURIZE(target, configuration)
-
-
-
-            # Export the frequencies we found for the clock-tree.
-
-            put_title('Clock-Tree')
-
-            for macro, expansion in justify(
-                (
-                    ('<', f'CLOCK_TREE_FREQUENCY_OF_{name}' ),
-                    ('>', f'{frequency:,}'.replace(',', "'")),
-                )
-                for name, frequency in tree
-                if name is not None
-            ):
-                Meta.define(macro, f'({expansion})')
-
-*/
